# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/contact/addFriend": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "添加好友"
                ],
                "summary": "添加好友",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "好友名称",
                        "name": "targetName",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/addGroup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "添加群组"
                ],
                "summary": "添加群组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "群id",
                        "name": "comId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/createGroup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "创建群组"
                ],
                "summary": "创建群组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "ownerId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "群头像",
                        "name": "icon",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/contact/findGroup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "查询群组"
                ],
                "summary": "查询群组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "ownerId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/index": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "跳转首页"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/toChat": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "跳转聊天页面"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/toRegister": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "跳转注册页面"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/FindByUserId": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "根据id查询用户"
                ],
                "summary": "根据id查询用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/createUser": {
            "post": {
                "tags": [
                    "注册"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/deleteUser": {
            "get": {
                "tags": [
                    "注销"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getMsgByRedis": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "查询redis的消息"
                ],
                "summary": "查询redis的消息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "userIdA",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "发消息着id",
                        "name": "userIdB",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "开始位置",
                        "name": "start",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "结束位置",
                        "name": "end",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "用户id",
                        "name": "isRev",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getUserList": {
            "get": {
                "tags": [
                    "获取用户列表"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "登录"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/searchFriend": {
            "post": {
                "tags": [
                    "查询好友"
                ],
                "summary": "查询好友",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/updateUser": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "修改用户"
                ],
                "summary": "修改用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "头像",
                        "name": "icon",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}